Fix watchtower signature
