+++ src/lib.rs
@@ -27,6 +27,8 @@ pub mod actors;
 
 pub mod tasks;
 
+pub mod utils;
+
 use git_version::git_version;
 
 const GIT_VERSION: &str = git_version!(fallback = "unknown");
